Update for embedded messages
Seriously unsure how safe this is but it looks correct to me.

Adhering to 10 embeds per message, splitting them up.

Looked back through the history, there is only ever 1 release per fork posted that I saw scrolling for as long as I did.

Removed "plugin header" as it is easy to see what plugin it is based on the fork hyperlink.

As inline on fields is silly and can't explicitly say 2x2x2, the no commit check needs to show a "N/A" to keep the alignment the same due to this annoyance.